requirements/developer.txt
-r production.txt

black==24.4.2
build==1.2.1
isort==5.13.2
pytest==8.3.2
pytest-cov==5.0.0
pytest-django==4.8.0
ruff==0.5.5